refactor

+++ app/src/main/java/it/manzolo/job/service/bluewatcher/activity/MainActivityFragment.kt
@@ -13,6 +13,7 @@ import it.manzolo.job.service.bluewatcher.utils.DateUtils
 import kotlinx.android.synthetic.main.fragment_main.*
 import java.io.File
 
+
 class MainActivityFragment : Fragment() {
 
     companion object {
@@ -26,16 +27,18 @@ class MainActivityFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        if (savedInstanceState == null) {
+// only create fragment if activity is started for the first time
+            val fileupdate = File(context?.cacheDir, "app.ava")
+            fileupdate.delete()
 
-        val fileupdate = File(context?.cacheDir, "app.ava")
-        fileupdate.delete()
-
-        //buttonUpdate.isEnabled = false
-        startUpdateService()
-        startJobService()
-        startWebsendService()
-        startLocationService()
-        startRebootService()
+            //buttonUpdate.isEnabled = false
+            startUpdateService()
+            startJobService()
+            startWebsendService()
+            startLocationService()
+            startRebootService()
+        }
 
 
     }
